Separate notes into notes and folders parts
